freelist: Prefix entries with memcheck redzone
Add redzone (default 16 bytes) before each entry of the freelist. The
redzone is marked as not accessible which allows memory access
checkers such as ASAN and Valgrind to detect out-of-bounds accesses
adjacent to the entries. In case on memory access checker is enabled,
the redzone size falls back to 0.

Use environment variable MEMCHECK_REDZONE_SIZE when configuring plugin
to control size of redzones.

Signed-off-by: Michael Axtmann <axtmannm@amazon.com>