freelist: Postfix reginfo_t with memcheck redzone
A memcheck redzone is added after the registration information
structure to detect out-of-bounce accesses at the border of the
registration information structure and the remaining freelist entry
that comes after that structure.

A redzone prefixing the reginfo_t structure is not necessary since the
freelist entry it starts with is already marked as not accessible.

Signed-off-by: Michael Axtmann <axtmannm@amazon.com>