freelist: Add fine-grained memcheck guards to allocated entries
Fine-grained memcheck guards are added to allocated freelist entries
to detect invalid accesses to allocated entries: The region before and
after registration information structure is set to undefined, the
element entry of the registration information structure is set to
noaccess, and the remainder of the registration information structure
is set to defined. In the case the "simple" freelist is used, the
whole entry is set to undefined.

Signed-off-by: Michael Axtmann <axtmannm@amazon.com>