API: Add neuron v5 plugin API providing RMA operations
Add plugin neuron API version v5. The API provides two major changes.

1. The API provides accept and connect functions that must not block
for the connection establishment, and instead should return
successfully with communicator == NULL with the expectation that is
will be called again until communicator != NULL.  This is compliant
with connection semantics introduced to NVIDIA APIs starting with
NVIDIA API v5.

2. The API extends API v4 by rkey extraction function `get_mr_key` and
RMA operations `iread()`, `iwrite()`, and `iwrite_inline()` to directly
write into remote memory regions via one-sided communication.
The RMA operations are only supported in case the `rma_supported` flag
in the device property is set to `1`. Data of `iwrite_inline()` may or
may not be inlined.
Note, this commit does not introduce implementations of the four new
API functions and sets `rma_supported` property to `0`.

Signed-off-by: Michael Axtmann <axtmannm@amazon.com>